requirements.txt
Django==4.2.23
django-simple-sso==0.7.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
